refactor: rename ScrollContainerContext to NavContext
The context only contains nav-related state (navLocked, navHidden).
Rename file, provider, and internal identifiers to match.
